feature(admin): Initial modal concept/mixin
You can call a modal like this: `this.$modal('title', 'body', data).$on(...)` where data will be passsed as vue data to the created modal which is outside of the root instance scope (for layering and dependency segregation).

Both title and body can be provided as a string which will be converted into dynamic components internally allowing you to use any passed data in the string. Or you can pass your own component for each of these and that will work too.